requirements.txt
asgiref==3.4.1
Django==3.2.2
djangorestframework==3.12.4
gunicorn==20.0.4
Pillow==9.0.0
psycopg2-binary==2.9.1
python-decouple==3.4
pytz==2021.3
sqlparse==0.4.2
